Fix undo of R_AARCH64_TLS_TPREL relocations.
prelink only does anything for these relocations in the ET_EXEC case.
Since the BFD linker writes the addend to the GOT for these
relocations in shared libraries, undo must not write a 0 for it to
recreate the original shared libraries.

This fixes the remaining test failures for AArch64, given glibc with
_dl_tlsdesc_undefweak hacked up as described by Samsung.

	* src/arch-aarch64.c (aarch64_undo_prelink_rela): Only write 0 for
	R_AARCH64_TLS_TPREL for ET_EXEC.

Signed-off-by: Joseph Myers <joseph@codesourcery.com>